requirements.txt
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
